# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API backend pour la marketplace Odin.",
        "title": "Marketplace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/catalog/agents": {
            "get": {
                "description": "Récupère la liste de tous les agents enregistrés",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Lister les agents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AgentListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Enregistre un nouvel agent sur la marketplace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Créer un agent",
                "parameters": [
                    {
                        "description": "Informations de l'agent",
                        "name": "agent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.SingleAgentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/catalog/agents/{id}": {
            "get": {
                "description": "Renvoie les détails d'un agent spécifique à partir de son UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Récupérer un agent par son ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID de l'agent",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SingleAgentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/packages/search": {
            "get": {
                "description": "Recherche un package par mot-clé (retourne une liste vide si aucun mot-clé fourni)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Rechercher des packages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Terme de recherche",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PackageSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CLIErrorResponse"
                        }
                    }
                }
            }
        },
        "/packages/{name}/{version}": {
            "get": {
                "description": "Récupère les informations d'un package et sa version pour le CLI",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Détail d'un package et sa version",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nom du package",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version semver ou 'latest'",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PackageDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.CLIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CLIErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AgentListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Agent"
                    }
                }
            }
        },
        "dto.CLIErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAgentRequest": {
            "type": "object",
            "required": [
                "agent_type",
                "category",
                "creator_id",
                "name",
                "runtime"
            ],
            "properties": {
                "agent_type": {
                    "type": "string",
                    "maxLength": 255
                },
                "category": {
                    "type": "string",
                    "maxLength": 255
                },
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "readme_markdown": {
                    "type": "string"
                },
                "runtime": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Le champ name est obligatoire"
                },
                "error": {
                    "type": "string",
                    "example": "Données invalides"
                }
            }
        },
        "dto.PackageDetailResponse": {
            "type": "object",
            "properties": {
                "agent": {
                    "$ref": "#/definitions/model.Agent"
                },
                "version": {
                    "$ref": "#/definitions/model.AgentVersion"
                }
            }
        },
        "dto.PackageSearchResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Agent"
                    }
                }
            }
        },
        "dto.SingleAgentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Agent"
                }
            }
        },
        "model.Agent": {
            "type": "object",
            "properties": {
                "agent_type": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "downloads_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_official_pick": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "readme_markdown": {
                    "type": "string"
                },
                "runtime": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.AgentVersion": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "type": "string"
                },
                "changelog": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_stable": {
                    "type": "boolean"
                },
                "manifest_json": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "security_run_id": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}